pyproject.toml
[project]
name = 'pycrawler'
version = '1.0.0'
dependencies = [
  "requests",
  "bs4",
  "pymongo",
  "mongoengine",
  "qdrant-client",
  "python-dateutil"
]

[tool.setuptools.packages.find]
where = ["."]  # list of folders that contain the packages (["."] by default)
include = ["*"]  # package names should match these glob patterns (["*"] by default)
exclude = []  # exclude packages matching these glob patterns (empty by default)
namespaces = false  # to disable scanning PEP 420 namespaces (true by default)